+++ RELEASENOTES.md
@@ -1,1 +1,1 @@
-
+- removed dead COPY from Docker files